chore(sonarlint): remove code flaw
had to do with int promotion and byte shifting, see https://sonarcloud.io/organizations/limechain/rules?open=java%3AS3034&rule_key=java%3AS3034

+++ src/main/java/com/limechain/trie/structure/nibble/NibblesToBytes.java
@@ -27,7 +27,7 @@ public class NibblesToBytes {
         for (int i = 0; i < halfLen; ++i) {
             Nibble n1 = nibbles.get(2 * i);
             Nibble n2 = nibbles.get(2 * i + 1);
-            byte b = (byte) ((n1.asByte() << 4) + n2.asByte());
+            byte b = (byte) ((n1.asInt() << 4) + n2.asInt());
             result.add(b);
         }
